requirements.txt
asgiref==3.2.10
Django==3.1.1
djangorestframework==3.12.1
pytz==2020.1
sqlparse==0.3.1
gunicorn==19.6.0
dj_database_url==0.5.0
psycopg2==2.7.5
django-environ==0.4.5